fix : Use GET request for query, move get request out from transaction search
