+++ src/app/api/posts/route.ts
@@ -57,7 +57,7 @@ export async function GET() {
 
     const postsWithImageUrl = await Promise.all(
       posts.map(async (post) => {
-        const url = await getPresignedUrl("book_images", post.book.coverImageKey);
+        const url = await getPresignedUrl("book_images", post.book.coverImagePath);
         return {
           ...post,
           book: {
