+++ src/app/api/transaction/route.ts
@@ -15,10 +15,25 @@ const createTransactionRequest = z.object({
   trackingURL: z.string(),
 })
 
+const beginningOfTime = new Date('0000-01-01T00:00:00Z');
+const endOfTime = new Date('9999-12-31T23:59:59Z');
+
+const getTransactionRequest = z.object({
+  userId: z.string().default(() => ""),
+  startDate: z.string().optional()
+    .refine((val) => !val || !isNaN(Date.parse(val)), { message: "Invalid date format" })
+    .transform((val) => (val == undefined ? beginningOfTime : new Date(val))),
+  endDate: z.string().optional()
+    .refine((val) => !val || !isNaN(Date.parse(val)), { message: "Invalid date format" })
+    .transform((val) => (val == undefined ? endOfTime : new Date(val))),
+  asBuyer: z.string().optional()
+    .transform((val) => (val == undefined ? true : val == "true")),
+  asSeller: z.string().optional()
+    .transform((val) => (val == undefined ? true : val == "true")),
+})
+
 export async function POST(req: NextRequest) {
   try {
-    console.log("json : ")
-    console.log(req)
     const parsedData = createTransactionRequest.safeParse(await req.json());
     if (!parsedData.success) {
       return NextResponse.json({ error: parsedData.error.errors }, { status: 400 });
@@ -58,20 +73,31 @@ export async function POST(req: NextRequest) {
   }
 }
 
-export async function GET() {
+export async function GET(req : NextRequest) {
   try {
-    const transactions = await prisma.transaction.findMany({
-      include: { 
-        buyer: true,
-        seller: true,
-        post: true,
-        failData: true 
-      },
-    });
+    const { searchParams } = new URL(req.url)
+    const queryParams = Object.fromEntries(searchParams.entries());
+    const parsedData = getTransactionRequest.safeParse(queryParams);
+    if (!parsedData.success) {
+      return NextResponse.json({ error: parsedData.error.errors }, { status: 400 });
+    }
 
-    return NextResponse.json(transactions);
+    const transactions = await prisma.transaction.findMany({
+      where: {
+        createOn: {
+          gte: parsedData.data.startDate,
+          lte: parsedData.data.endDate,
+        },
+        ...(parsedData.data.userId !== '' ? { OR: [
+          (parsedData.data.asBuyer) ? { buyerId: parsedData.data.userId } : {},
+          (parsedData.data.asSeller) ? { sellerId: parsedData.data.userId } : {},
+        ].filter(Boolean) } : {}),
+      }
+    })
+
+    return NextResponse.json(transactions, { status: 201 });
   } catch (error) {
     if (error instanceof Error) console.error("Error getting transactions", error.stack);
-    return NextResponse.json({ error: "Cannot get transactions" }, { status: 500 });
+    return NextResponse.json({ error: "Cannot get a transaction" }, { status: 500 });
   }
 }
